# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Was It Worth It? — reflection-based discretionary spending tracker
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PurchaseStatus =
  (typeof PurchaseStatus)[keyof typeof PurchaseStatus];

export const PurchaseStatus = {
  waiting: "waiting",
  reflected: "reflected",
} as const;

export type Verdict = (typeof Verdict)[keyof typeof Verdict];

export const Verdict = {
  worth_it: "worth_it",
  not_worth_it: "not_worth_it",
  uncertain: "uncertain",
} as const;

export type Category = (typeof Category)[keyof typeof Category];

export const Category = {
  food: "food",
  clothing: "clothing",
  entertainment: "entertainment",
  subscriptions: "subscriptions",
  travel: "travel",
  other: "other",
} as const;

export interface Purchase {
  id: string;
  name: string;
  amount: number;
  category: Category;
  purchaseDate: string;
  status: PurchaseStatus;
  readyForReflection: boolean;
  readyAt?: string;
  createdAt: string;
}

export interface Reflection {
  id: string;
  purchaseId: string;
  verdict: Verdict;
  comments?: string;
  reflectedAt: string;
}

export interface PurchaseDetail {
  purchase: Purchase;
  reflection?: Reflection;
}

export interface CreatePurchaseInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 0 */
  amount: number;
  category: Category;
  purchaseDate: string;
}

export interface UpdatePurchaseInput {
  /** @minLength 1 */
  name?: string;
  /** @minimum 0 */
  amount?: number;
  category?: Category;
  purchaseDate?: string;
}

export interface SubmitReflectionInput {
  verdict: Verdict;
  comments?: string;
}

export interface DashboardSummary {
  readyForReflectionCount: number;
  totalPurchases: number;
  totalReflected: number;
  nextPending: Purchase;
}

export interface CategoryInsight {
  category: Category;
  totalReflected: number;
  worthItCount: number;
  notWorthItCount: number;
  uncertainCount: number;
  worthItPercent: number;
  totalSpent: number;
}

export interface Insights {
  categories: CategoryInsight[];
  overallWorthItPercent: number;
  totalSpent: number;
  regretSpent: number;
}

export interface Profile {
  /**
   * @minimum 1
   * @maximum 90
   */
  reflectionWindowDays: number;
  emailReminders: boolean;
  inAppReminders: boolean;
}

export interface UpdateProfileInput {
  /**
   * @minimum 1
   * @maximum 90
   */
  reflectionWindowDays?: number;
  emailReminders?: boolean;
  inAppReminders?: boolean;
}

export type ListPurchasesParams = {
  status?: PurchaseStatus;
  search?: string;
  category?: Category;
};