# lib/api-zod/src/generated/types/categoryInsight.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Was It Worth It? — reflection-based discretionary spending tracker
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";

export interface CategoryInsight {
  category: Category;
  totalReflected: number;
  worthItCount: number;
  notWorthItCount: number;
  uncertainCount: number;
  worthItPercent: number;
  totalSpent: number;
}