# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Was It Worth It? — reflection-based discretionary spending tracker
 * OpenAPI spec version: 0.1.0
 */
import type { Purchase } from "./purchase";

export interface DashboardSummary {
  readyForReflectionCount: number;
  totalPurchases: number;
  totalReflected: number;
  nextPending: Purchase;
}