# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Was It Worth It? — reflection-based discretionary spending tracker
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  /**
   * @minimum 1
   * @maximum 90
   */
  reflectionWindowDays: number;
  emailReminders: boolean;
  inAppReminders: boolean;
}