# lib/api-zod/src/generated/types/purchase.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Was It Worth It? — reflection-based discretionary spending tracker
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";
import type { PurchaseStatus } from "./purchaseStatus";

export interface Purchase {
  id: string;
  name: string;
  amount: number;
  category: Category;
  purchaseDate: Date;
  status: PurchaseStatus;
  readyForReflection: boolean;
  readyAt?: Date;
  createdAt: Date;
}