# lib/api-zod/src/generated/types/purchaseDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Was It Worth It? — reflection-based discretionary spending tracker
 * OpenAPI spec version: 0.1.0
 */
import type { Purchase } from "./purchase";
import type { Reflection } from "./reflection";

export interface PurchaseDetail {
  purchase: Purchase;
  reflection?: Reflection;
}