# lib/api-zod/src/generated/types/reflection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Was It Worth It? — reflection-based discretionary spending tracker
 * OpenAPI spec version: 0.1.0
 */
import type { Verdict } from "./verdict";

export interface Reflection {
  id: string;
  purchaseId: string;
  verdict: Verdict;
  comments?: string;
  reflectedAt: Date;
}