# lib/api-zod/src/generated/types/submitReflectionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Was It Worth It? — reflection-based discretionary spending tracker
 * OpenAPI spec version: 0.1.0
 */
import type { Verdict } from "./verdict";

export interface SubmitReflectionInput {
  verdict: Verdict;
  comments?: string;
}