# lib/api-zod/src/generated/types/updatePurchaseInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Was It Worth It? — reflection-based discretionary spending tracker
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";

export interface UpdatePurchaseInput {
  /** @minLength 1 */
  name?: string;
  /** @minimum 0 */
  amount?: number;
  category?: Category;
  purchaseDate?: Date;
}