# lib/api-zod/src/generated/types/verdict.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Was It Worth It? — reflection-based discretionary spending tracker
 * OpenAPI spec version: 0.1.0
 */

export type Verdict = (typeof Verdict)[keyof typeof Verdict];

export const Verdict = {
  worth_it: "worth_it",
  not_worth_it: "not_worth_it",
  uncertain: "uncertain",
} as const;